# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Forex Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SignalSignal = (typeof SignalSignal)[keyof typeof SignalSignal];

export const SignalSignal = {
  BUY: "BUY",
  SELL: "SELL",
} as const;

export type SignalStructureType =
  (typeof SignalStructureType)[keyof typeof SignalStructureType];

export const SignalStructureType = {
  BOS: "BOS",
  CHOCH: "CHOCH",
  NONE: "NONE",
} as const;

export type SignalTrend = (typeof SignalTrend)[keyof typeof SignalTrend];

export const SignalTrend = {
  BULLISH: "BULLISH",
  BEARISH: "BEARISH",
  NEUTRAL: "NEUTRAL",
} as const;

export type SignalStatus = (typeof SignalStatus)[keyof typeof SignalStatus];

export const SignalStatus = {
  ACTIVE: "ACTIVE",
  HIT_TP: "HIT_TP",
  HIT_SL: "HIT_SL",
  EXPIRED: "EXPIRED",
} as const;

export interface Signal {
  id: number;
  pair: string;
  signal: SignalSignal;
  timeframe: string;
  entry: number;
  stopLoss: number;
  takeProfit: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  confidenceScore: number;
  reasons: string[];
  structureType: SignalStructureType;
  trend: SignalTrend;
  hasOrderBlock: boolean;
  hasSupportResistance: boolean;
  riskRewardRatio: number;
  status: SignalStatus;
  createdAt: string;
}

export type CreateSignalBodySignal =
  (typeof CreateSignalBodySignal)[keyof typeof CreateSignalBodySignal];

export const CreateSignalBodySignal = {
  BUY: "BUY",
  SELL: "SELL",
} as const;

export type CreateSignalBodyStructureType =
  (typeof CreateSignalBodyStructureType)[keyof typeof CreateSignalBodyStructureType];

export const CreateSignalBodyStructureType = {
  BOS: "BOS",
  CHOCH: "CHOCH",
  NONE: "NONE",
} as const;

export type CreateSignalBodyTrend =
  (typeof CreateSignalBodyTrend)[keyof typeof CreateSignalBodyTrend];

export const CreateSignalBodyTrend = {
  BULLISH: "BULLISH",
  BEARISH: "BEARISH",
  NEUTRAL: "NEUTRAL",
} as const;

export interface CreateSignalBody {
  pair: string;
  signal: CreateSignalBodySignal;
  timeframe: string;
  entry: number;
  stopLoss: number;
  takeProfit: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  confidenceScore: number;
  reasons: string[];
  structureType: CreateSignalBodyStructureType;
  trend: CreateSignalBodyTrend;
  hasOrderBlock: boolean;
  hasSupportResistance: boolean;
  riskRewardRatio: number;
}

export interface AnalyzeRequest {
  pair: string;
  timeframe: string;
}

export type AnalysisResultSignal =
  (typeof AnalysisResultSignal)[keyof typeof AnalysisResultSignal];

export const AnalysisResultSignal = {
  BUY: "BUY",
  SELL: "SELL",
  NEUTRAL: "NEUTRAL",
} as const;

export type AnalysisResultStructureType =
  (typeof AnalysisResultStructureType)[keyof typeof AnalysisResultStructureType];

export const AnalysisResultStructureType = {
  BOS: "BOS",
  CHOCH: "CHOCH",
  NONE: "NONE",
} as const;

export type AnalysisResultTrend =
  (typeof AnalysisResultTrend)[keyof typeof AnalysisResultTrend];

export const AnalysisResultTrend = {
  BULLISH: "BULLISH",
  BEARISH: "BEARISH",
  NEUTRAL: "NEUTRAL",
} as const;

export type AnalysisResultSupportZone = {
  high: number;
  low: number;
};

export type AnalysisResultResistanceZone = {
  high: number;
  low: number;
};

export type AnalysisResultOrderBlockZoneType =
  (typeof AnalysisResultOrderBlockZoneType)[keyof typeof AnalysisResultOrderBlockZoneType];

export const AnalysisResultOrderBlockZoneType = {
  BULLISH: "BULLISH",
  BEARISH: "BEARISH",
} as const;

export type AnalysisResultOrderBlockZone = {
  high: number;
  low: number;
  type: AnalysisResultOrderBlockZoneType;
} | null;

export interface AnalysisResult {
  pair: string;
  timeframe: string;
  signal: AnalysisResultSignal;
  entry: number;
  stopLoss: number;
  takeProfit: number;
  confidenceScore: number;
  reasons: string[];
  structureType: AnalysisResultStructureType;
  trend: AnalysisResultTrend;
  hasOrderBlock: boolean;
  hasSupportResistance: boolean;
  riskRewardRatio: number;
  supportZone: AnalysisResultSupportZone;
  resistanceZone: AnalysisResultResistanceZone;
  orderBlockZone?: AnalysisResultOrderBlockZone;
}

export type DashboardSummaryTopPairsItem = {
  pair: string;
  count: number;
  winRate: number;
};

export interface DashboardSummary {
  totalSignals: number;
  activeSignals: number;
  winRate: number;
  avgConfidence: number;
  buySignals: number;
  sellSignals: number;
  topPairs: DashboardSummaryTopPairsItem[];
  recentActivity: Signal[];
}

export interface ForexPair {
  symbol: string;
  base: string;
  quote: string;
  pip: number;
  decimals: number;
}

export type ListSignalsParams = {
  pair?: string;
  signal?: ListSignalsSignal;
  timeframe?: string;
};

export type ListSignalsSignal =
  (typeof ListSignalsSignal)[keyof typeof ListSignalsSignal];

export const ListSignalsSignal = {
  BUY: "BUY",
  SELL: "SELL",
} as const;