# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Forex Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all trading signals
 */
export const ListSignalsQueryParams = zod.object({
  pair: zod.coerce.string().optional(),
  signal: zod.enum(["BUY", "SELL"]).optional(),
  timeframe: zod.coerce.string().optional(),
});

export const listSignalsResponseConfidenceScoreMin = 0;
export const listSignalsResponseConfidenceScoreMax = 100;

export const ListSignalsResponseItem = zod.object({
  id: zod.number(),
  pair: zod.string(),
  signal: zod.enum(["BUY", "SELL"]),
  timeframe: zod.string(),
  entry: zod.number(),
  stopLoss: zod.number(),
  takeProfit: zod.number(),
  confidenceScore: zod
    .number()
    .min(listSignalsResponseConfidenceScoreMin)
    .max(listSignalsResponseConfidenceScoreMax),
  reasons: zod.array(zod.string()),
  structureType: zod.enum(["BOS", "CHOCH", "NONE"]),
  trend: zod.enum(["BULLISH", "BEARISH", "NEUTRAL"]),
  hasOrderBlock: zod.boolean(),
  hasSupportResistance: zod.boolean(),
  riskRewardRatio: zod.number(),
  status: zod.enum(["ACTIVE", "HIT_TP", "HIT_SL", "EXPIRED"]),
  createdAt: zod.coerce.date(),
});
export const ListSignalsResponse = zod.array(ListSignalsResponseItem);

/**
 * @summary Create a new trading signal
 */
export const createSignalBodyConfidenceScoreMin = 0;
export const createSignalBodyConfidenceScoreMax = 100;

export const CreateSignalBody = zod.object({
  pair: zod.string(),
  signal: zod.enum(["BUY", "SELL"]),
  timeframe: zod.string(),
  entry: zod.number(),
  stopLoss: zod.number(),
  takeProfit: zod.number(),
  confidenceScore: zod
    .number()
    .min(createSignalBodyConfidenceScoreMin)
    .max(createSignalBodyConfidenceScoreMax),
  reasons: zod.array(zod.string()),
  structureType: zod.enum(["BOS", "CHOCH", "NONE"]),
  trend: zod.enum(["BULLISH", "BEARISH", "NEUTRAL"]),
  hasOrderBlock: zod.boolean(),
  hasSupportResistance: zod.boolean(),
  riskRewardRatio: zod.number(),
});

/**
 * @summary Get a specific signal
 */
export const GetSignalParams = zod.object({
  id: zod.coerce.number(),
});

export const getSignalResponseConfidenceScoreMin = 0;
export const getSignalResponseConfidenceScoreMax = 100;

export const GetSignalResponse = zod.object({
  id: zod.number(),
  pair: zod.string(),
  signal: zod.enum(["BUY", "SELL"]),
  timeframe: zod.string(),
  entry: zod.number(),
  stopLoss: zod.number(),
  takeProfit: zod.number(),
  confidenceScore: zod
    .number()
    .min(getSignalResponseConfidenceScoreMin)
    .max(getSignalResponseConfidenceScoreMax),
  reasons: zod.array(zod.string()),
  structureType: zod.enum(["BOS", "CHOCH", "NONE"]),
  trend: zod.enum(["BULLISH", "BEARISH", "NEUTRAL"]),
  hasOrderBlock: zod.boolean(),
  hasSupportResistance: zod.boolean(),
  riskRewardRatio: zod.number(),
  status: zod.enum(["ACTIVE", "HIT_TP", "HIT_SL", "EXPIRED"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a signal
 */
export const DeleteSignalParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Run live analysis on a forex pair
 */
export const AnalyzeSignalBody = zod.object({
  pair: zod.string(),
  timeframe: zod.string(),
});

export const AnalyzeSignalResponse = zod.object({
  pair: zod.string(),
  timeframe: zod.string(),
  signal: zod.enum(["BUY", "SELL", "NEUTRAL"]),
  entry: zod.number(),
  stopLoss: zod.number(),
  takeProfit: zod.number(),
  confidenceScore: zod.number(),
  reasons: zod.array(zod.string()),
  structureType: zod.enum(["BOS", "CHOCH", "NONE"]),
  trend: zod.enum(["BULLISH", "BEARISH", "NEUTRAL"]),
  hasOrderBlock: zod.boolean(),
  hasSupportResistance: zod.boolean(),
  riskRewardRatio: zod.number(),
  supportZone: zod.object({
    high: zod.number(),
    low: zod.number(),
  }),
  resistanceZone: zod.object({
    high: zod.number(),
    low: zod.number(),
  }),
  orderBlockZone: zod
    .object({
      high: zod.number(),
      low: zod.number(),
      type: zod.enum(["BULLISH", "BEARISH"]),
    })
    .nullish(),
});

/**
 * @summary Dashboard statistics summary
 */
export const getDashboardSummaryResponseRecentActivityItemConfidenceScoreMin = 0;
export const getDashboardSummaryResponseRecentActivityItemConfidenceScoreMax = 100;

export const GetDashboardSummaryResponse = zod.object({
  totalSignals: zod.number(),
  activeSignals: zod.number(),
  winRate: zod.number(),
  avgConfidence: zod.number(),
  buySignals: zod.number(),
  sellSignals: zod.number(),
  topPairs: zod.array(
    zod.object({
      pair: zod.string(),
      count: zod.number(),
      winRate: zod.number(),
    }),
  ),
  recentActivity: zod.array(
    zod.object({
      id: zod.number(),
      pair: zod.string(),
      signal: zod.enum(["BUY", "SELL"]),
      timeframe: zod.string(),
      entry: zod.number(),
      stopLoss: zod.number(),
      takeProfit: zod.number(),
      confidenceScore: zod
        .number()
        .min(getDashboardSummaryResponseRecentActivityItemConfidenceScoreMin)
        .max(getDashboardSummaryResponseRecentActivityItemConfidenceScoreMax),
      reasons: zod.array(zod.string()),
      structureType: zod.enum(["BOS", "CHOCH", "NONE"]),
      trend: zod.enum(["BULLISH", "BEARISH", "NEUTRAL"]),
      hasOrderBlock: zod.boolean(),
      hasSupportResistance: zod.boolean(),
      riskRewardRatio: zod.number(),
      status: zod.enum(["ACTIVE", "HIT_TP", "HIT_SL", "EXPIRED"]),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary List supported forex pairs
 */
export const ListPairsResponseItem = zod.object({
  symbol: zod.string(),
  base: zod.string(),
  quote: zod.string(),
  pip: zod.number(),
  decimals: zod.number(),
});
export const ListPairsResponse = zod.array(ListPairsResponseItem);