# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Forex Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultOrderBlockZone } from "./analysisResultOrderBlockZone";
import type { AnalysisResultResistanceZone } from "./analysisResultResistanceZone";
import type { AnalysisResultSignal } from "./analysisResultSignal";
import type { AnalysisResultStructureType } from "./analysisResultStructureType";
import type { AnalysisResultSupportZone } from "./analysisResultSupportZone";
import type { AnalysisResultTrend } from "./analysisResultTrend";

export interface AnalysisResult {
  pair: string;
  timeframe: string;
  signal: AnalysisResultSignal;
  entry: number;
  stopLoss: number;
  takeProfit: number;
  confidenceScore: number;
  reasons: string[];
  structureType: AnalysisResultStructureType;
  trend: AnalysisResultTrend;
  hasOrderBlock: boolean;
  hasSupportResistance: boolean;
  riskRewardRatio: number;
  supportZone: AnalysisResultSupportZone;
  resistanceZone: AnalysisResultResistanceZone;
  orderBlockZone?: AnalysisResultOrderBlockZone;
}