# lib/api-zod/src/generated/types/createSignalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Forex Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSignalBodySignal } from "./createSignalBodySignal";
import type { CreateSignalBodyStructureType } from "./createSignalBodyStructureType";
import type { CreateSignalBodyTrend } from "./createSignalBodyTrend";

export interface CreateSignalBody {
  pair: string;
  signal: CreateSignalBodySignal;
  timeframe: string;
  entry: number;
  stopLoss: number;
  takeProfit: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  confidenceScore: number;
  reasons: string[];
  structureType: CreateSignalBodyStructureType;
  trend: CreateSignalBodyTrend;
  hasOrderBlock: boolean;
  hasSupportResistance: boolean;
  riskRewardRatio: number;
}