# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Forex Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryTopPairsItem } from "./dashboardSummaryTopPairsItem";
import type { Signal } from "./signal";

export interface DashboardSummary {
  totalSignals: number;
  activeSignals: number;
  winRate: number;
  avgConfidence: number;
  buySignals: number;
  sellSignals: number;
  topPairs: DashboardSummaryTopPairsItem[];
  recentActivity: Signal[];
}